backend/pandora/requirement.txt
wheel==0.36.2
django==3.2
djangorestframework==3.12.4
# django-rest-swagger==2.2.0
drf-yasg==1.20.0
django-filter==2.4.0
django-cors-headers==3.7.0
python-redis-lock[django]==3.7.0
django-redis==4.12.1
django-mptt==0.12.0
celery==5.0.5
django-rest-auth==0.9.5
redis-collections==0.9.0
pillow==8.2.0
xlrd==2.0.1
openpyxl==3.0.7
qrcode==6.1
gevent==21.1.2
sphinx==3.5.3

django_silk==4.1.0
psycopg2==2.8.6
msgpack==1.0.2
umsgpack==0.1.0
piexif==1.1.3
mysqlclient==2.0.3
django-mysql==3.11.1
demjson==2.2.4
mongoengine==0.23.0
ldap3==2.9
django-picklefield==3.0.1
ruamel.yaml==0.17.4
retry==0.9.2
protobuf==3.15.8
gunicorn==20.1.0
pyjwkest==2.25.1
pyyaml==5.4.1
# tensorflow
# uwsgi==2.0.18


